Implement a BaseLayout for the home and forum screens to share the implementation for the header and bottom navigation
