+++ mobile/Quizzard/screens/BaseLayout.tsx
@@ -1,0 +1,88 @@
+// BaseLayout.tsx
+import React from 'react';
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
+const BaseLayout = ({ children, navigation }) => {
+  const handleLogout = () => {
+    // Simulate logout and navigate back to the login screen
+    navigation.navigate('Login');
+  };
+
+  return (
+    <View style={styles.container}>
+      {/* Header Section */}
+      <View style={styles.header}>
+        <Text style={styles.appName}>Quizzard</Text>
+        <View style={styles.icons}>
+          <TouchableOpacity style={styles.iconButton}>
+            <Ionicons name="person-outline" size={24} color="black" />
+          </TouchableOpacity>
+          <TouchableOpacity style={styles.iconButton} onPress={handleLogout}>
+            <Ionicons name="power-outline" size={24} color="black" />
+          </TouchableOpacity>
+        </View>
+      </View>
+
+      {/* Dynamic Content Section */}
+      <View style={styles.body}>
+        {children}
+      </View>
+
+      {/* Bottom Navigation Bar */}
+      <View style={styles.navBar}>
+        <TouchableOpacity style={styles.navButton} onPress={() => navigation.navigate('Home')}>
+          <Ionicons name="home" size={28} color="black" />
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.navButton} onPress={() => navigation.navigate('Forum')}>
+          <Ionicons name="chatbox-outline" size={28} color="black" />
+        </TouchableOpacity>
+      </View>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+  },
+  header: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: 16,
+    backgroundColor: '#fff',
+  },
+  appName: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    color: '#6a0dad', // Dark purple color for the app name
+  },
+  icons: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  iconButton: {
+    marginLeft: 16,
+  },
+  body: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  navBar: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingVertical: 12,
+    paddingHorizontal: 40,
+    backgroundColor: '#f2f2f2',
+    borderTopWidth: 1,
+    borderColor: '#ccc',
+  },
+  navButton: {
+    padding: 10,
+  },
+});
+
+export default BaseLayout;
