+++ mobile/Quizzard/screens/ForumScreen.tsx
@@ -1,33 +1,22 @@
+// RightPage.tsx
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { Text, StyleSheet } from 'react-native';
+import BaseLayout from './BaseLayout';
 
-const ForumScreen = () => {
+const ForumScreen = ({ navigation }) => {
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Forum</Text>
-      <Text style={styles.subtitle}>Welcome to the forum!</Text>
-    </View>
+    <BaseLayout navigation={navigation}>
+      <Text style={styles.title}>Welcome to the forum!</Text>
+    </BaseLayout>
   );
 };
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 16,
-    backgroundColor: '#fff',
-  },
   title: {
     fontSize: 32,
     fontWeight: 'bold',
     color: '#6a0dad',
     marginBottom: 16,
-  },
-  subtitle: {
-    fontSize: 20,
-    color: '#6a0dad',
-    marginBottom: 40,
   },
 });
